refactor(backend): edited paths in seeder file
